Add code to load weather data and log to console
I got the basics setup. The main.js file will call the loadPage function.
At the momement, loadPage calls the getWeatherData function with my zip
code. This will call the Geocode API to get the latitude and longitute,
and then it will call the OpenWeather API to retreive the weather data.
Once the weather data is loaded it calls the displaWeatherData function.
For now, this logs the data to the console. Later, it will show the
weahter on the screen.