Make weather display on the page
I added the code to the displayWeatherData function to update the
elements with the required data. I also added a variable to store the
zipcode that the user inputs in the weather-data-handler file. This is
used to refresh the weather data (refreshWeatherData) when the user
presses the refresh button. It will call the getWeatherData function
using the stored zipcode. The zipcode will change when the user enters
a new one.