Add Pexel API integration
So now during the weather data load process it will use the data
retrieved to determine the users state. It will take the state and send
it over to the image-handler which will use that information to run a
search using the Pexel API. Once some images are found, it will randomly
pick an image from the first page and then display it to the user.